+++ src/RelationshipMatrices.jl
@@ -15,6 +15,7 @@ include("grm-bits.jl")
 include("hinv.jl")
 include("irm.jl")
 
-export nrm, nrm_diag, Ainv, ainv, grm, kinship, validate_pedigree, hinv, Hinv
+export nrm, nrm_diag, Ainv, ainv, grm, irm, irm_locus, kinship,
+    validate_pedigree, hinv, Hinv
 
 end # module RelationshipMatrices